Added deaths to command handler.
