+++ index.js
@@ -72,81 +72,6 @@ client.on('message', async msg => {
 
 
   if (
-		msg.content.toLowerCase().startsWith(config.discordPrefix + 'punishments')
-	) {
-		try {
-			let response = await fetch(
-				config.apiUrl + '/mc/punishment/latest?limit=10'
-			);
-			let punishments = await response.json();
-			let punMsg = [];
-			var punType;
-			punishments.forEach(punishment => {
-				if (!punishment.punisherLoaded)
-					punishment.punisherLoaded = { name: 'Console' };
-				time = new Date(punishment.issued).toString().split(' ')[4];
-				const difference = punishment.expires - punishment.issued;
-				let duration;
-				if (difference === 0) duration = null;
-				else if (punishment.expires === -1) duration = 'permanent';
-				else
-					duration = humanize(punishment.expires - punishment.issued, {
-						largest: 1,
-					});
-				punishment.type.toLowerCase() === 'warn'
-					? (punType = 'warned')
-					: punType;
-				punishment.type.toLowerCase() === 'mute'
-					? (punType = 'muted')
-					: punType;
-				punishment.type.toLowerCase() === 'ban'
-					? (punType = 'banned')
-					: punType;
-				punishment.type.toLowerCase() === 'kick'
-					? (punType = 'kicked')
-					: punType;
-				punMsg.push(
-					`🔹 \`${time}\` **${
-						punishment.punishedLoaded.name
-					}** was ${punType} for **${punishment.reason}** ${
-						duration ? '(' + duration + ')' : ''
-					}`
-				);
-			});
-			let embed = new MessageEmbed();
-			embed.setTitle(`10 most recent punishments on ${config.servername}`);
-			embed.setColor('RED');
-			embed.setDescription(punMsg.join('\n'));
-			msg.channel.send({ embed });
-		} catch (err) {
-			msg.channel.send('An error occurred.\n\n\n```js\n' + err + '```');
-		}
-	} else if (
-		msg.content.toLowerCase().startsWith(config.discordPrefix + 'deaths')
-	) {
-		try {
-			let response = await fetch(config.apiUrl + '/mc/death/latest');
-			let deaths = await response.json();
-			var deathMsg = [];
-			deaths.forEach(death => {
-				if (!death.killerLoaded) death.killerLoaded = { name: 'Environment' };
-				time = new Date(death.date).toString().split(' ')[4];
-				deathMsg.push(
-					`🔹 \`${time}\` **${death.killerLoaded.name}** killed **${death.playerLoaded.name}** with **${death.killerItem}**`
-				);
-			});
-			let embed = new MessageEmbed();
-			embed.setTitle(`4 most recent deaths on ${config.servername}`);
-			embed.setColor('RED');
-			embed.setDescription(deathMsg.join('\n'));
-			embed.setFooter(
-				'Due to API limitations, I can only display up to 4 recent deaths :('
-			);
-			msg.channel.send({ embed });
-		} catch (err) {
-			msg.channel.send('An error occurred.\n\n\n```js\n' + err + '```');
-		}
-	} else if (
 		msg.content.toLowerCase().startsWith(config.discordPrefix + 'server')
 	) {
 		if (!args[0]) return msg.channel.send('You need to provide a server type');
